+++ packages/react/src/components/TextInput/index.tsx
@@ -3,12 +3,13 @@ import { Input, Prefix, TextInputContainer } from './styles'
 
 export interface TextInputProps extends ComponentProps<typeof Input> {
   prefix?: string
+  inputSize?: 'sm' | 'md'
 }
 
 export const TextInput = forwardRef<ElementRef<typeof Input>, TextInputProps>(
-  ({ prefix, ...props }: TextInputProps, ref) => {
+  ({ prefix, inputSize = 'md', ...props }: TextInputProps, ref) => {
     return (
-      <TextInputContainer>
+      <TextInputContainer inputSize={inputSize}>
         {!!prefix && <Prefix>{prefix}</Prefix>}
         <Input ref={ref} {...props} />
       </TextInputContainer>