+++ packages/react/src/components/TextInput/styles.ts
@@ -10,7 +10,7 @@ export const TextInputContainer = styled('div', {
   alignItems: 'center',
 
   variants: {
-    size: {
+    inputSize: {
       sm: {
         padding: '$2 $3',
       },
@@ -30,7 +30,7 @@ export const TextInputContainer = styled('div', {
   },
 
   defaultVariants: {
-    size: 'md',
+    inputSize: 'md',
   },
 })