+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ntn-ui/tokens
 
+## 3.0.0
+
+### Major Changes
+
+- adjust property inputSize in textInput component
+
 ## 2.0.0
 
 ### Major Changes